+++ Jogo/Player.h
@@ -22,7 +22,7 @@ private:
 	SwordAttack* swordHitbox;
 
 public:
-	Player(GraphicsManager* graphicsManager, float* dt, int id, float spriteScale, sf::Vector2f position, sf::Vector2f bodySize, std::string pathToTexture, std::string textureName, float speed, int hp);
+	Player(GraphicsManager* graphicsManager, float* dt, int id, float spriteScale, sf::Vector2f position, sf::Vector2f bodySize, std::string pathToTexture, std::string textureName, float speed, long int hp);
 	Player();
 	~Player();
